Development

+++ Project/frontend/src/pages/managecus/index.js
@@ -1,6 +1,7 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React, { useEffect, useState, useRef } from "react";
 import ReactToPrint from 'react-to-print';
+import Swal from 'sweetalert2'
 
 import { Link } from "react-router-dom";
 import "./managecus.css";
@@ -132,12 +133,47 @@ export default function CusManagement() {
 
     // This method will delete a record
     async function deleteRecord(id) {
-        await fetch(`${process.env.REACT_APP_BACKEND_URL}/loyalty/delete/${id}`, {
-            method: "DELETE"
-        });
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+                confirmButton: 'text-white bg-gradient-to-r from-green-400 via-green-500 to-green-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-green-300 dark:focus:ring-green-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2',
+                cancelButton: 'text-white bg-gradient-to-r from-red-400 via-red-500 to-red-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-red-300 dark:focus:ring-red-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2'
+            },
+            buttonsStyling: false
+        })
 
-        const newRecords = records.filter((el) => el._id !== id);
-        setRecords3(newRecords);
+        swalWithBootstrapButtons.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this! You will the Delete the This Customer Loyalty Level Details!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`${process.env.REACT_APP_BACKEND_URL}/loyalty/delete/${id}`, {
+                    method: "DELETE"
+                });
+
+                const newRecords = records.filter((el) => el._id !== id);
+                setRecords3(newRecords);
+
+                swalWithBootstrapButtons.fire(
+                    'Deleted!',
+                    'This Customer Loyalty Level has been deleted.',
+                    'success'
+                )
+            } else if (
+                /* Read more about handling dismissals below */
+                result.dismiss === Swal.DismissReason.cancel
+            ) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelled',
+                    'This Customer Loyalty Level is Not Deleted:)',
+                    'info'
+                )
+            }
+        })
     }
 
     // This method will map out the records on the table
